C++ refactoring: to and from json (#1164)
* delete temporary refactoring.h header

* remove Content from json

* cleanup

* add namesapace

* use unordered_map

* Preserve order in v1 RecordForm.contents.

* revert changes

* add py36 check

* disable test for python < 3.6

* disable tests for python < 3.6

* disable tests for python < 3.6

* disable tests for python < 3.6

* skip tests that require python 3.6

* add check for py27

* move common classes to a common header

* cleanup

+++ src/libawkward/builder/RecordBuilder.cpp
@@ -71,7 +71,7 @@ namespace awkward {
 
     std::stringstream out;
     out << "{\"class\": \"RecordArray\", \"contents\": {";
-    for (int64_t i = 0;  i < contents_.size();  i++) {
+    for (size_t i = 0;  i < contents_.size();  i++) {
       if (i != 0) {
         out << ", ";
       }
@@ -83,6 +83,7 @@ namespace awkward {
       out << "\"parameters\": {\"__record__\": " + util::quote(name_) + "}, ";
     }
     out << "\"form_key\": \"" + form_key.str() + "\"}";
+
     return out.str();
   }
